requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.27.2
websockets==13.1
python-dotenv